Move errors segment modals
